Add style to the landing page
Followed the layout given by the assignment and made the page like it.
I didn't use any images for now, in their place I created divs to
represent the space an image would take. Also couldn't change the font
style of a paragraph to light, for some reason. But mostly it matches
the given layout pretty nicely.